+++ R/complete_ae_data.R
@@ -41,52 +41,52 @@
                           missing_text = "Unknown", missing_location = "first") {
 
   # check inputs ---------------------------------------------------------------
-  if (is.null(by) && !is.null(by_values))
+  if (is_empty(by) && !is_empty(by_values))
     stop("Cannot specify `by_values=` without also specifying `by=`.", call. = FALSE)
-  if (!is.null(by) && inherits(data[[by]], "factor") && !is.null(by_values))
+  if (!is_empty(by) && inherits(data[[by]], "factor") && !is_empty(by_values))
     stop("Cannot specify `by_values=` when `by=` is a factor as it is expected
          that factor levels contain all possible `by_values`.", call. = FALSE)
   if (!rlang::is_string(missing_text)) {
     stop("The `missing_text=` argument must be a string.", call. = FALSE)
   }
-  if (!is.null(by_values) && !is.character(by_values)) {
+  if (!is_empty(by_values) && !is.character(by_values)) {
     stop("The `by_values=` argument must be a character vector.", call. = FALSE)
   }
 
   # check ID between data and id_df
   # Check the id and strata column names match `data=`
-  if (!is.null(id_df) && !(id %in% names(id_df))) {
+  if (!is_empty(id_df) && !(id %in% names(id_df))) {
     stop("The `id=` column must be present in `id_df=`.", call. = FALSE)
   }
-  if (!is.null(id_df) && !is.null(strata) && !(strata %in% names(id_df))) {
+  if (!is_empty(id_df) && !is_empty(strata) && !(strata %in% names(id_df))) {
     stop("The `strata=` column must be present in `id_df=`.", call. = FALSE)
   }
 
   # Check the id and strata type matches between `data=` and `id_df`
-  if (!is.null(id_df) && !identical(class(data[[id]]), class(id_df[[id]]))) {
+  if (!is_empty(id_df) && !identical(class(data[[id]]), class(id_df[[id]]))) {
     stop("The class of the `id=` column must match in both `data=` and `id_df=`.", call. = FALSE)
   }
-  if (!is.null(id_df) && !is.null(strata) && !identical(class(data[[strata]]), class(id_df[[strata]]))) {
+  if (!is_empty(id_df) && !is_empty(strata) && !identical(class(data[[strata]]), class(id_df[[strata]]))) {
     stop("The class of the `strata=` column must match in both `data=` and `id_df=`.", call. = FALSE)
   }
 
   # Check the id and strata columns are not missing
   if (any(is.na(data[[id]])) ||
-      ( !is.null(strata) && any(is.na(data[[strata]]))) ) {
+      ( !is_empty(strata) && any(is.na(data[[strata]]))) ) {
     stop("Columns `id=` and `strata=` cannot be missing in `data=`", call. = FALSE)
   }
   # Check the id and strata columns are not missing
-  if (!is.null(id_df) &&
+  if (!is_empty(id_df) &&
       (any(is.na(id_df[[id]])) ||
-      ( !is.null(strata) && any(is.na(id_df[[strata]])))) ) {
+      ( !is_empty(strata) && any(is.na(id_df[[strata]])))) ) {
     stop("Columns `id=` and `strata=` cannot be missing in `id_df=`", call. = FALSE)
   }
 
   # 3. Check all ID/strata combos appear in `data=`
-  if (!is.null(id_df) && any(duplicated(id_df[c(id, strata)]))) {
+  if (!is_empty(id_df) && any(duplicated(id_df[c(id, strata)]))) {
     stop("Disallowed duplicate `id=`/`strata=` combinations found in `id_df=`.", call. = FALSE)
   }
-  if (!is.null(id_df) &&
+  if (!is_empty(id_df) &&
       nrow(dplyr::anti_join(dplyr::distinct(data[c(id, strata)]),
                             id_df[c(id, strata)],
                             by = c(id, strata))) > 0) {
@@ -94,17 +94,17 @@
   }
 
   # check for missing soc
-  if (!is.null(soc) && any(is.na(data[soc]))) {
+  if (!is_empty(soc) && any(is.na(data[soc]))) {
     stop("At least one `soc` is missing.", call. = FALSE)
   }
 
   # check for missing ae
-  if (!is.null(ae) && any(is.na(data[ae]))) {
+  if (!is_empty(ae) && any(is.na(data[ae]))) {
     stop("At least one `ae` is missing.", call. = FALSE)
   }
 
   # check for an ae listed under more than one soc
-  if (!is.null(soc) &&
+  if (!is_empty(soc) &&
       nrow(dplyr::distinct(data[c(ae, soc)])) != nrow(dplyr::distinct(data[ae]))) {
     stop("The `ae` levels must be unique across all `soc` levels.", call. = FALSE)
   }
@@ -133,7 +133,7 @@
     )
 
   # if data frame of ids is supplied, add IDs obs to data ----------------------
-  if (!is.null(id_df)) {
+  if (!is_empty(id_df)) {
     data <-
       id_df %>%
       select(!!!lst_name_recode[c("id", "strata")]) %>%
@@ -167,7 +167,7 @@
   }
 
   # identifying rows that will be used in tabulation ---------------------------
-  if (!is.null(soc)) {
+  if (!is_empty(soc)) {
     data_full <-
       data_full %>%
       arrange(across(any_of(c("id", "strata", "soc", "by")))) %>%
@@ -204,29 +204,29 @@
 }
 
 .prepare_by_levels <- function(data, by, by_values, initial_missing, initial_dummy) {
-  if (!is.null(by) && initial_dummy %in% data[["by"]]) {
+  if (!is_empty(by) && initial_dummy %in% data[["by"]]) {
     stringr::str_glue("Level '{initial_dummy}' cannot ",
                       "appear in the levels of the `by=` variable.") %>%
       stop(call. = FALSE)
   }
-  if (!is.null(by) && initial_missing %in% data[["by"]] && any(is.na(data[["by"]]))) {
+  if (!is_empty(by) && initial_missing %in% data[["by"]] && any(is.na(data[["by"]]))) {
     stringr::str_glue("Level '{initial_missing}' cannot appear in the levels ",
                       "of the `by=` variable when missing data present.") %>%
       stop(call. = FALSE)
   }
 
-  if (!is.null(by_values) && initial_dummy %in% by_values) {
+  if (!is_empty(by_values) && initial_dummy %in% by_values) {
     stringr::str_glue("Level '{initial_dummy}' cannot ",
                       "appear in the levels of the `by_values=` argument") %>%
       stop(call. = FALSE)
   }
-  if (!is.null(by_values) && initial_missing %in% by_values && any(is.na(data[["by"]]))) {
+  if (!is_empty(by_values) && initial_missing %in% by_values && any(is.na(data[["by"]]))) {
     stringr::str_glue("Level '{initial_missing}' cannot appear in the levels ",
                       "of the `by_values=` argument when missing data present.") %>%
       stop(call. = FALSE)
   }
 
-  if (is.null(by)) {
+  if (is_empty(by)) {
     data$by <- factor("Overall")
   }
 
@@ -234,7 +234,7 @@
     data$by <- factor(data$by)
   }
 
-  if (!is.null(by_values)) {
+  if (!is_empty(by_values)) {
     if (!rlang::is_empty(setdiff(levels(data$by), by_values))) {
       stop("All levels of `by=` variable must appear in  `by_values=`",
            call. = FALSE)