+++ R/gtreg-package.R
@@ -1,8 +1,8 @@
 #' @keywords internal
 #' @importFrom purrr %||%
-#' @importFrom rlang .env .data sym is_missing :=
+#' @importFrom rlang .env .data sym is_missing := is_empty
 #' @importFrom dplyr mutate select filter ungroup group_by arrange across
-#' @importFrom broom.helpers .formula_list_to_named_list .select_to_varnames
+#' @importFrom cli cli_alert_danger cli_code cli_alert_info
 "_PACKAGE"
 
 # allowing for the use of the dot when piping